scraper/requirements.txt
# Foreclosure Lead Scraper Requirements
# For R740 Server Deployment

# Core scraping
crawl4ai>=0.3.0
playwright>=1.48.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selectolax>=0.3.0
parsel>=1.8.0

# HTTP clients
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Browser automation
selenium>=4.25.0
undetected-chromedriver>=3.5.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Queue and scheduling
redis>=5.0.0
celery>=5.3.0
apscheduler>=3.10.0

# Data processing
pandas>=2.2.0
pydantic>=2.6.0
python-dateutil>=2.8.0

# Email automation
aiosmtplib>=3.0.0
email-validator>=2.1.0
jinja2>=3.1.0

# Proxy management
proxy-pool>=2.0.0

# OCR for scanned documents
pytesseract>=0.3.0
pdf2image>=1.17.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Utilities
tenacity>=8.2.0
python-dotenv>=1.0.0
structlog>=24.1.0
rich>=13.7.0

# Monitoring
prometheus-client>=0.19.0
sentry-sdk>=1.39.0
